LINE_app/requirements.txt
pytz==2018.9
Pillow==8.0.1
requests==2.21.0
streamlit==0.84.0